Set execution constraints by target.
Provides a new flag --experimental_add_exec_constraints_to_target that allows setting execution constraints on targets in addition to those statically declared on them.

Note that this flag should not be necessary: all execution constraints should be derivable from the target, the rule it is instantiating and available toolchains it depends on. The flag is nonetheless introduced (hopefully) in a temporary fashion to allow working around issues where execution platform and strategy do not match, as discussed in https://docs.google.com/document/d/1U9HzdDmtRnm244CaRM6JV-q2408mbNODAMewcGjnnbM. If and when that design is implemented this flag will be removed again. Use at your own risk.

RELNOTES: None.
PiperOrigin-RevId: 278367121