C++: Add grep_incudes to Bazel cc_common.link()
Even though grep_includes is not available in Bazel. We want to make the API
exactly the same internally and externally.

create_linking_context_from_compilation_outputs was already had this attribute
in Bazel.

In a separate CL make sure we don't get an error when grep_includes is needed but the binary wasn't passed.

RELNOTES:none
PiperOrigin-RevId: 278369911