Make CcToolchainConfigInfo.proto account for the legacy features patching logic: https://docs.bazel.build/versions/1.1.0/cc-toolchain-config-reference.html#legacy-features-patching-logic
Before this change, CcToolchainConfigInfo.proto would generate a prototext based on what the toolchain maintainer has specified in their Starlark rule, and would not account for the transformations we do with our features patching logic.
Now that we don't support CROSSTOOL prototext within Bazel, and the CcToolchainConfigInfo.proto is intended to be used in comparison scripts, we can make the .proto method return a CToolchain prototext that contains all the added features, action_configs, artifact_name_patterns, in the order they appear after Bazel reshuffles them.

This also allows us to lazily generate the proto, which is also done in this change.

RELNOTES: None.
PiperOrigin-RevId: 278582045