Async + ActionFileSystem: don't throw NPE
When using async execution in combination with an action file system,
the try-with-resources block in ActionExecutionFunction causes the
ActionFileSystem instance to be cleared before the action is done
running.

This change basically replaces the try-with-resources block with a
manual implementation and does not close the ActionExecutionContext (and
associated ActionFileSystem) if we are in a continuation case.

The proper solution is to rewrite ActionExecutionFunction to enforce
correct use of continutations in all cases and drop the distinction
between async and non-async execution at this level.

PiperOrigin-RevId: 278840442