Use weak-referenced ConcurrentMap to prevent memory leaks.
Previously, the concurrent maps in ArtifactNestedSetFunction have strong
references to their keys. Therefore, when an Artifact's SkyKey is no longer
valid and is removed from Skyframe's map (e.g. when a file is removed from the
dependency list), its entry still stays on artifactToSkyValueMap, preventing it
from being garbage collected, which results in a memory leak.

By having the references being weak, we can prevent this memory leak.

Note that this also changes the map's behavior, as it now uses identity instead
of the Object#equals method to compare keys. This is mostly fine because:
- The SkyKeys aren't repeatedly instantiated for the same Artifact.
- This doesn't pollute Skyframe, since the weakmap is completely separated from
  Skyframe's internal map.

RELNOTES: None
PiperOrigin-RevId: 278847339