bazel packages: don't return TreeMaps from native.existing_rule
This change does not include a regression test because it would be
oddly specific to a scenario that will likely never regress again.
The test we really want is a precondition check in all functions like
SkylarkDict.add that causes them to reject non-Starlark values.
However, that much bigger change---or series of changes---must wait
for a follow-up.

PiperOrigin-RevId: 279130469