Don't intern the intermediate String used for starlark String -> bazel Label conversions.
This String is parsed and split during the Label conversion process, so interning this String (the *original* one) is useless and we're just needlessly paying the CPU and memory cost of the Interner. Note that Label already does its own interning of the "name" part (for "//a:b", the name will be "b"); that _is_ theoretically useful.

Also don't bother calling STRING.convert since it's just useless method overhead. Label#convert already does an instanceof check.

RELNOTES: None
PiperOrigin-RevId: 279145542