Refactor remote package to only have one remote cache impl.
Historically the remote package shipped two implementations
for remote caching: GrpcRemoteCache and SimpleBlobStoreActionCache.
Today there is no good reason to keep this structure and the
duplication of code and tests that comes with it. This is the
final refactoring of a long series:

 * 5fc91a75082d5015654160d43c168e5eb74fe2ae
 * 12ebb84a7645baffdbaf9f40cbfcec996a540ffe
 * 8ac6bdcf36950f1dba5e7c7edc61ac53f490e3f1
 * 9fb83b4869c2569b5cd7af488631e950c8a3e4fd
 * 36611c380323ca91986543d4970743bdf0a0efbd
 * a6b5b0540a0167f84437127d3dd6cef804286d6f
 * d40933d479dfb8350f5f8d236296e18064714210
 * 797f2925e4aafddca520984a474b97c16eceb9f0

This change makes it so that there is one RemoteCache that
has a RemoteCacheClient. The RemoteCacheClient has one
implementation per caching protocol:

 * DiskCacheClient (formerly OnDiskBlobStore)
 * GrpcCacheClient (formerly GrpcRemoteCache)
 * HttpCacheClient (formerly HttpBlobStore)
 * CombinedDiskHttpCacheClient (formerly CombinedDiskHttpBlobStore)

A single RemoteCache type for all protocols will allow
composition of caching protocols. In particular, grpc and disk
caching. Additionally, this change will allow us to fix a longstanding
bug where for the HTTP protocol uploads of outputs for an action are
not done in parallel.