bazel syntax: rename SkylarkList.Tuple to Tuple
Tuple is one of the core data types of Starlark.
SkylarkList, which should be named Sequence, is the interface common to all indexable sequences.

Also, remove is SkylarkList.isTuple method.
The only class that isTuple is Tuple.

By a minor miracle, there were no references to Tuple from within the
copybara code base, so this renaming is unusually straightforward.

PiperOrigin-RevId: 279730954