bazel syntax: eliminate all but one calls to Runtime.setupStarlarkFunctions
This functionality has moved to class Starlark, in two variants
that reflect the two different intentions of their callers:
- addModule adds a single name to the environment (e.g. attr, native)
- addMethods adds all the methods of a value to the environment (e.g. glob, exports_files, etc)

setupStarlarkFunctions must remain in place temporarily until the last
remaining use, in the copybara code base, is removed. This is trivial but
cannot be done in the same change.

Also, make these CallUtils methods private:
 getMethods
 getMethodNames (1 overload)
 getBuiltinCallable

PiperOrigin-RevId: 279746790