Don't throw when encountering providers with multiple "direct" resources/assets
Exports should be bundled as siblings of the library that exports them, so
it's always possible for "direct resources/assets" to not be a singleton list.

This works around the fact that AARs are an implicit output of every
android_library rule, even though creating an AAR for a rule with 'exports'
would be nonsensical.

PiperOrigin-RevId: 279796125