Enable combined disk cache for grpc caching protocol.
This change enables the combined disk cache for the gRPC
protocol too.

$ bazel build src:bazel --disk_cache=/path/to/dir
--remote_cache=grpcs://...

It does not yet also enable it for remote execution as
this will need a bit more work. In particular it will
require some refactorings around RemoteCacheClient#findMissingBlobs
as when used with remote execution findMissingBlobs() should
only check the remote cache and not the local disk cache. Else,
it could happen that an input file is missing from the local
cache but not from the remote cache in which case Bazel would
attempt to upload the file again.

RELNOTES: The --disk_cache flag can now also be used together
with the gRPC remote cache.