Refactor the Ninja variable value class, and correct some lexer/parser issues.
Ninja variable value now contains the list of functions for composing the expanded value.
For the text fragments, the function return the text literal, for variable reference, it uses the passed variable expander function to obtain the variable value.

Corrected lexing/parsing problems:
- In text fragment, do not treat = and | symbols as special.
- Do not allow the single $ symbol in text (always require it to be escaped).
- In a variable value, allow non-escaped : symbol, except for the case in build line (this is still not coded).
- Do not try to cut out the escape symbol; rather, un-escape text in parser. For convenience, introduce one more NinjaToken: escaped text.

All the lexing/parsing changes are done in correspondence with Ninja manual and source code.

Closes #10227.

PiperOrigin-RevId: 280404778