Execute sandboxfs reconfigurations in parallel.
I tried adding unit tests for this but doing so is very complicated in
their current form.  Given that this feature is still very experimental
and not in wide use, I'll postpone doing that until I'm able to prove
its worth.

For now, I ran a few builds by hand using the still-unreleased sandboxfs
0.2.0 with parallel support and confirmed this to work.

RELNOTES: None.
PiperOrigin-RevId: 280566455