Remove SourceManifestAction.Builder
This Builder class is superfluous - it simply wraps Runfiles.Builder,
and forwards all calls to that. Instead, update all callers to construct
a Runfiles object directly and pass that in.

This is in preparation for changing SymlinkTreeAction to also take a
Runfiles object instead of reading the manifest back in from disk. This
makes it possible to remove the manifests if they aren't needed, as well
as fixing the handling of whitespace in runfiles paths.

PiperOrigin-RevId: 281018764