SymlinkTreeAction: require a Runfiles object
This is not particularly useful in itself, but updates the callers to
match the new expectation, and also points out the one place (filesets)
where we may need to do additional work.

I'll make use of the Runfiles reference in a subsequent change to avoid
reading back the local file into Bazel.

PiperOrigin-RevId: 281263527