Add interface and implementation that will allow the remote module to inject remote execution functionality into repository_rule.
1) Introduce the RepositoryRemoteExecutor interface which will be
implemented by the RemoteModule in a follow up change. The intent
is for the RemoteModule to register a RepositoryRemoteExecutorFactory
which the BazelRepositoryModule calls before each command in order
to obtain a RepositoryRemoteExecutor object. This extra layer of
indirection is necessary because remote execution can be enabled
and disabled between commands without causing reanalysis.

2) Add logic to repository_ctx.execute() to execute a command remotely
using RepositoryRemoteExecutor

RELNOTES: None
PiperOrigin-RevId: 281279763