Proto v2 for aquery proto output formats (--incompatible_proto_output_v2)
To try out this change, please include --incompatible_proto_output_v2 in your aquery invocation. Note that this flag is experimental and its API will change in the future.

The existing analysis.proto uses String as the format for messages' ids. This follows the convention, but results in a bloated output size. This CL introduces the 2nd version of the proto (analysis_v2.proto), which uses uint64 as ids instead. We've observed a significant improvement in output size.

RELNOTES: --incompatible_proto_output_v2: proto v2 for aquery proto output formats, which reduces the output size compared to v1. Note that the messages' ids in v2 are in uint64 instead of string like in v1.
PiperOrigin-RevId: 281709375