Make "blaze test <alias target>" run the test.
In order to stay consistent, also fix the tests() query operator. Not binaries(), though: query logic is kinda complicated so I don't want to shave that particular yak. Implementing ConfiguredTargetValueAccessor#getPrerequisites() was enough yak shaving for the day.

A drive-by cleanup that I *could* *have* done, but didn't, is to allow aliases in test_suite.tests and maybe make it configurable: now the logic is all there, so the only change that would be required is to actually allow aliases where we now allow only tests and test_suite rules and remove the non-configurable bit from the attribute. If we want that, it should be done in a separate change because it's separate functionality.

Taking all the select() branches for alias.actual is a questionable approach, but we don't have enough data where we (currently) compute the resolution of target patterns to do better and it's consistent with "blaze query", so I guess that'll do for now. I'll file that under the "blaze query should really be aware of configurations" label.

RELNOTES: None.
PiperOrigin-RevId: 281730608