Invalidate the analysis cache if a starlark option changes.
Without this change it was possible for an action to be in the skyframe cache whose outputs' contents had changed based on a starlark option without invalidating the action. Then if the action was requested again Bazel would assume that its outputs were fine and not re-run the action.

RELNOTES: None.
PiperOrigin-RevId: 281962111