bazel syntax: rationalize Depset constructors
This change renames the various functions formerly called 'of'.

Overloading is fine for two methods that apply same logic to different
representations of their input (e.g. int vs long, CharSequence vs
String), but it is an abuse to use it for different algorithms.

Also, remove unnecessary Location parameters. (Locations are not
needed for EvalExceptions as they are filled in by the call
machinery.)

PiperOrigin-RevId: 282004067