bazel syntax: reintroduce element validity check to depset contructor
Also, declare concrete subclasses of analysis.ConfiguredTarget as @Immutable,
and thus Starlark-hashable.

It is possible that user code passes values of other classes that fail
the element validity check. In that case we must either withdraw the check
and phase it in using a flag, or, consider whether the specific value should
be declared immutable/hashable, as was appropriate for ConfiguredTarget.

PiperOrigin-RevId: 282015280