Implemented a new event called DiscoveredInputsEvent that gets posted if the action has discoveredInputs in the ActionExecutionFunction. The discovered inputs event was added to tell the CriticalPathComputer that it took n amount of time to go through the discover input phase and to categorize it at parse time.
With the introduction of this new event, the CriticalPathComponent is now created earlier in the CriticalPathComputer#outputArtifactToComponent map so it opens up the chances of the race condition mentioned in b/19612679. To avoid the chances of hitting the race condition (regarding shared actions) the following was implemented:

Default the CriticalPathComponent#isRunning flag to false and only flip it to true when ActionStartedEvent occurs. The three scenarios where it would be flipped to true are
   1. A new CriticalPathComponent is created and should start running.
   2. A CriticalPathComponent has been created through discover inputs and is starting
      to execute.
   3. An action was rewound and starts again.

RELNOTES: None.
PiperOrigin-RevId: 282021149