bazel: rename SkylarkValue
This change moves skylarkinterface.SkylarkValue to syntax.StarlarkValue,
and also skylarkinterface.SkylarkPrinter to syntax.Printer, where its
functionality is merged into the existing class of that name.

This change is entirely mechanical, except for syntax.Printer (naturally),
and MethodLibrary and SkylarkEvaluationTest, which needed a tweak
due the removal of the Printer.debugPrint static method.

All classes related to the representation of values are now in the evaluator package;
skylarkinterface now contains only the annotations.

This is a breaking change for copybara.

BEGIN_PUBLIC
bazel: rename SkylarkValue
END_PUBLIC
PiperOrigin-RevId: 282410278