Symlink creation: pass symlinks to OutputService
This allows implementations of the OutputService to skip reading the
manifest file and instead use the in-memory data.

This is the second attempt for https://github.com/bazelbuild/bazel/commit/83b70caca5952e75489b43cae21837c047a475af including a fix and tests
for null artifacts.

The reason for the breakage is that Python runfiles use null to indicate
that we should create empty files (in addition to the symlinks). However,
this was not caught due to insufficient test coverage for the case of
using an active OutputService.

PiperOrigin-RevId: 282526946