Be explicit about the queue being used for memory pressure, sleep and suspend events. Add debug logging to make manual testing of anomalies easier (memory pressure and sleep anomalies can't be unit tested very effectively).
Turns out that IONotificationPortSetDispatchQueue returns silently doing nothing if it is passed a NULL queue, and apparently when I migrated the code over to using IONotificationPortSetDispatchQueue and started using DISPATCH_TARGET_QUEUE_DEFAULT (which is equivalent to NULL) I didn't test the actual sleep functionality sufficiently to verify that it was working (focusing on the SIGNAL functionality that did work correctly with the DISPATCH_TARGET_QUEUE_DEFAULT). Using a single queue for both the memory pressure, suspend and sleep events allows us to verify that the queue is being initialized correctly, and gives us some minimal debugging functionality by naming the queue for us.

RELNOTES: none
PiperOrigin-RevId: 282575977