bazel syntax: put depset element check behind a flag
This change adds a flag, --incompatible_check_elements_in_all_depset_constructors,
that causes the depset(direct=...) constructor to check the validity of its elements,
which it should have done all along but, regrettably, did not.

Validity is currently defined as "immutable", but will eventually become "hashable",
but that's another incompatible change.

PiperOrigin-RevId: 282633198