Add $USE_BAZEL_VERSION and .bazelversion support to Bazel's wrapper script.
Also simplify the installation instructions for Ubuntu and how to

RELNOTES: Bazel's Debian package and the binary installer now include an improved wrapper that understands `<WORKSPACE>/.bazelversion` files and the `$USE_BAZEL_VERSION` environment variable. This is similar to what Bazelisk offers (https://github.com/bazelbuild/bazelisk#how-does-bazelisk-know-which-bazel-version-to-run-and-where-to-get-it-from), except that it works offline and integrates with apt-get.
PiperOrigin-RevId: 283031899