bazel syntax: support struct.to_{json,proto} bound method
github.com/bazelbuild/bazel/pull/8931 was an incomplete fix to b/21392896:
it allowed 'x.f()' to be expressed as 'y=x.f; y()' in most cases, but not all.
In particular, it did not address the to_json and to_proto methods of
structs (which we should get rid of, BTW). This change is an ad-hoc fix for
those two methods. A more principled fix will have to wait till an integrated
clean-up of Info/StructImpl/NativeInfo.

This change is a prerequisite for fixing github.com/bazelbuild/bazel/issues/10339,
which is in turn a prerequisite for various simplifications and optimizations to
Starlark's function calling mechanism.

I audited Google's .bzl corpus for cases where the three-argument
getattr(x, name, dflt) was called with either an arbitrary name or
a name obtained from dir(x) without filtering, and where subsequent
operations might change their behavior if given a bound method instead
of 'dflt'. This turned up only two locations, both easily fixed (CL 283224061).

PiperOrigin-RevId: 283433923