VFS: add RootedPathAndCasing
This class stores a RootedPath and the String
value of its relative part.

This class allows distinguishing RootedPaths on
case-insensitive filesystems, if the RootedPaths
refer to the same file but use different casing.

The class only compares the casing of the relative
parts, not of the root parts. Reason is, this
class will be used to validate the casing of
package paths, which are always relative to the
package root whose casing we needn't worry about.

See https://github.com/bazelbuild/bazel/issues/8799

Change-Id: Ie742f95a7feb3935c93cbba6db9278b63cfcc63a