A build's or test's critical path run time was surpassing the execution run time, which is impossible. The root cause was that within the same critical path, multiple actions were run in parallel. Therefore the total run time is the union of all those actions instead of the aggregate (which was the original implementation).
For example, there are some cases where an action's dependency can run and complete during the time it's dependent action is in its discover inputs phase. In this scenario the total run time would be the
  "dependent's finish - start time"
rather than:
  "dependent's finish - start time" + "dependency's finish - start time".

The solution was to create a timeline of each component that has it's child's timeline. Since each child is requested and finished before the start of the dependent component, we can assure that the dependent calling the child's timeline returns the child's longest and finalized timeline. The timeline of type RangeSet was also used because this can keep track of any gaps in between each time range, keeping this logic consistent with the original implementation.

RELNOTES: Critical path run time should not have a longer run time than total execution run time.
PiperOrigin-RevId: 283535419