New SkyFunction: PathCasingLookupFunction
The new PathCasingLookupFunction is a SkyFunction
that can validate the casing of a path.

On case-insensitive filesystems if "foo/BUILD"
exists then "FOO/BUILD" also exists (because they
mean the same file), but this should not imply
that the "//foo" and "//FOO" packages both exist.

The new SkyFunction can validate whether
"FOO/BUILD" is the correct casing. The casing is
correct if the file's name in its parent
directory's listing is exactly the same
(case-sensitively) as the expected name.

In a later PR we can change the
PackageLookupFunction to deny loading if the path
is not correctly cased.

Change-Id: Ibb2a5f4f6c6041bdfc537b33d0c6fc55e568b998
Motivation: https://github.com/bazelbuild/bazel/issues/8799