Stream aquery proto outputs.
aquery outputs proto results in a monolithic manner, which means storing the
entire ActionGraphContainer in memory before printing everything to an output
file. This could cause OOM error on larger queries.

This CL makes aquery stream results instead. This requires a change in
proto.ActionGraphContainer: instead of a wrapper object like before, it's now a
collection of ActionGraphComponent. ActionGraphComponent would have a
Discriminator, similar to query's build.Target.

This is a breaking change for users of --incompatible_proto_output_v2.

RELNOTES: None
PiperOrigin-RevId: 283726774